+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
 - do not hijack "enter" i.e. let the browser do the job and create new paragraphs
+- add multiple add, remove and update paragraphs at once in the "undo-redo" stack
 
 # 0.0.2 (2022-02-04)